refactor(email): Create nodemailer package
