+++ apps/docs/content/docs/dev/email/nodemailer.mdx
@@ -1,0 +1,71 @@
+---
+title: Nodemailer (SMTP)
+description: Send emails using SMTP with the Nodemailer adapter.
+---
+
+| Cloud            | Self-Hosted  | Links                                                   |
+| ---------------- | ------------ | ------------------------------------------------------- |
+| ❌ Not Supported | ✅ Supported | [NPM Package](https://www.npmjs.com/package/nodemailer) |
+
+## Usage
+
+<Steps>
+<Step>
+### Installation
+
+import { Tab, Tabs } from "fumadocs-ui/components/tabs";
+
+<Tabs groupId='package-manager' persist items={['bun', 'pnpm', 'npm']} label='Install nodemailer adapter'>
+
+```bash tab="bun"
+bun i @vitnode/nodemailer -D
+```
+
+```bash tab="pnpm"
+pnpm i @vitnode/nodemailer -D
+```
+
+```bash tab="npm"
+npm i @vitnode/nodemailer -D
+```
+
+</Tabs>
+
+</Step>
+<Step>
+### Import the adapter
+
+```ts title="vitnode.api.config.ts"
+// [!code ++]
+import { NodemailerEmailAdapter } from "@vitnode/nodemailer";
+import { buildApiConfig } from "@vitnode/core/vitnode.config";
+
+export const vitNodeApiConfig = buildApiConfig({
+  email: {
+    // [!code ++:6]
+    adapter: NodemailerEmailAdapter({
+      from: process.env.NODE_MAILER_FROM,
+      host: process.env.NODE_MAILER_HOST,
+      password: process.env.NODE_MAILER_PASSWORD,
+      user: process.env.NOD_EMAILER_USER,
+    }),
+  },
+});
+```
+
+</Step>
+<Step>
+
+### Environment Variables
+
+Add the following environment variables to your `.env` file:
+
+```bash title=".env"
+NODE_MAILER_FROM=xxx
+NODE_MAILER_HOST=xxx
+NODE_MAILER_PASSWORD=xxx
+NOD_EMAILER_USER=xxx
+```
+
+</Step>
+</Steps>
