--- apps/docs/content/docs/dev/email/smtp.mdx
@@ -1,13 +1,0 @@
----
-title: SMTP
-description: SMTP configuration for sending emails
----
-
-<Callout type="warn" title="Only Self-Hosted supported!">
-  This documentation is for self-hosted VitNode instances only. You cannot use
-  this if you are planning to deploy your application to the cloud.
-</Callout>
-
-<Callout type="warn" title="This documentation is under construction! 🚧">
-  We're working hard to bring you the best documentation experience.
-</Callout>
