+++ packages/nodemailer/.swcrc
