+++ packages/nodemailer/README.md
@@ -1,0 +1,20 @@
+# (VitNode) Nodemailer (SMTP) Adapter
+
+This package provides a Nodemailer email adapter for VitNode, enabling email sending capabilities using SMTP.
+
+<p align="center">
+  <br>
+  <a href="https://vitnode.com/" target="_blank">
+    <picture>
+      <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/VitNode/vitnode/canary/assets/logo/vitnode_logo_dark.svg">
+      <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/VitNode/vitnode/canary/assets/logo/vitnode_logo_light.svg">
+      <img alt="VitNode Logo" src="https://raw.githubusercontent.com/VitNode/vitnode/canary/assets/logo/vitnode_logo_light.svg" width="400">
+    </picture>
+  </a>
+  <br>
+  <br>
+</p>
+
+| Cloud            | Self-Hosted  | Links                                                   | Documentation                                         |
+| ---------------- | ------------ | ------------------------------------------------------- | ----------------------------------------------------- |
+| ❌ Not Supported | ✅ Supported | [NPM Package](https://www.npmjs.com/package/nodemailer) | [Docs](https://vitnode.com/docs/dev/email/nodemailer) |
