+++ packages/nodemailer/eslint.config.mjs
@@ -1,5 +1,4 @@
 import eslintVitNode from "@vitnode/config/eslint";
-import eslintVitNodeReact from "@vitnode/config/eslint.react";
 import { fileURLToPath } from "node:url";
 import { dirname } from "node:path";
 
@@ -7,7 +6,6 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
 
 export default [
   ...eslintVitNode,
-  ...eslintVitNodeReact,
   {
     languageOptions: {
       parserOptions: {
