+++ plugins/blog/eslint.config.mjs
@@ -1,4 +1,5 @@
 import eslintVitNode from "@vitnode/config/eslint";
+import eslintVitNodeReact from "@vitnode/config/eslint.react";
 import { fileURLToPath } from "node:url";
 import { dirname } from "node:path";
 
@@ -6,6 +7,7 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
 
 export default [
   ...eslintVitNode,
+  ...eslintVitNodeReact,
   {
     languageOptions: {
       parserOptions: {
